Tree level as cipher nonces, change URN to erisx3
Use the tree level of a block as nonce for the Chacha20 cipher.
This prevents the plaintext of a leaf block being handled as key
and refrence pairs if the block level is wrong.

This is a breaking change, so replace "erisx2" with "erisx3" in
URNs.

Add a test to check that ErisIngest works properly.