Refactor to use dedicated get/put state objects
Replace use of FuturVar[T] with the FutureBlock and the FutureGet
and FuturePut variants. These objects contain the parameters and
state required for get and put operations as well as a stack of
callbacks for completetion. This refactor was intended to improve
performance by reducing copying but yielding more useful things
like ensuring blocks are verified and makes composite stores
easier to implement.